Renaming forecasting to prdiction
