+++ tests/forecast/test_dataset_characteristics.py
@@ -1,4 +1,4 @@
-from turbo_ml.forecast.dataset_characteristics import StatisticalParametersExtractor, DatasetDescription
+from turbo_ml.model_prediction.dataset_characteristics import StatisticalParametersExtractor, DatasetDescription
 from datasets import get_iris, get_breast_cancer
