+++ tests/test_imports.py
@@ -2,7 +2,7 @@ def test_basic_imports():
     from turbo_ml import __name__
     from turbo_ml.algorithms import __name__
     from turbo_ml.base import __name__
-    from turbo_ml.forecast import __name__
+    from turbo_ml.model_prediction import __name__
     from turbo_ml.utils import __name__
     from turbo_ml.interface import __name__
     from turbo_ml.preprocessing import __name__
