+++ turbo_ml/meta_model/load_and_dump.py
@@ -1,5 +1,5 @@
 import torch
-from turbo_ml.model_preparation.as_meta_model import Best_Model
+from turbo_ml.meta_model.as_meta_model import Best_Model
 model = Best_Model(15, 36)
 model.load_state_dict(torch.load('model.pth'))
 model = model.to('cpu')
