+++ turbo_ml/meta_model/meta_model_search.py
@@ -1,9 +1,9 @@
 import os
 from functools import cache
-from ..forecast.model_forecast import Forecast
+from ..model_prediction.model_prediction import Predictor
 from ..preprocessing import sota_preprocessor
 from ..base import __ALL_MODELS__
-from ..forecast.dataset_characteristics import DatasetDescription
+from ..model_prediction.dataset_characteristics import DatasetDescription
 import pandas as pd
 import torch
 import torch.nn as nn
@@ -28,7 +28,7 @@ class Best_Model(nn.Module):
         return x
 
 
-class MetaModelGuesser(Forecast):
+class MetaModelGuesser(Predictor):
     """ Search for the best meta model for a given dataset and model """
 
     def __init__(self, device='cpu'):
