+++ turbo_ml/model_prediction/__init__.py
@@ -1,3 +1,3 @@
-from .model_forecast import Forecast, ExhaustiveSearch
+from .model_prediction import Predictor, ExhaustiveSearch
 from .dataset_characteristics import StatisticalParametersExtractor
 from ..meta_model.meta_model_search import MetaModelGuesser
